requirements.txt
decorator==4.1.2
dj-database-url==0.3.0
Django==1.8.4
django-cors-headers==2.1.0
django-environ==0.4.4
django-extensions==1.5.6
django-filter==0.11.0
djangorestframework==3.2.3
dotenv==0.0.5
enum34==1.1.6
gunicorn==19.3.0
ipython==4.0.0
ipython-genutils==0.2.0
Markdown==2.6.2
newrelic==2.54.0.41
pathlib2==2.3.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==2.9.0
psycopg2==2.6.1
ptyprocess==0.5.2
pytz==2015.4
scandir==1.6
simplegeneric==0.8.1
six==1.9.0
traitlets==4.3.2
Werkzeug==0.10.4
whitenoise==2.0.3
